==> code/1_concat_all_data.ipynb <==
import pandas as pd
import os
import sys
import time
from tqdm import *
import pickle
from collections import Counter, defaultdict

def save_variable(v,filename):
    f=open(filename,'wb')
    pickle.dump(v,f)
    f.close()
    return filename
 
def load_variavle(filename):
    f=open(filename,'rb')
    r=pickle.load(f)
    f.close()
    return r

path = '../data/data_all/'
id_map = dict(enumerate(os.listdir(path)))
new_map = {v:k for k, v in id_map.items()}

start = time.clock()
for i, j in tqdm(enumerate(os.listdir(path))):
    data = pd.read_csv(path+j)
    data['sample_file_name'] = int(i)
    if i == 0:
        data.to_csv('../data/data_all.csv', index=False)
    else:
        data.to_csv('../data/data_all.csv', index=False, header=False, mode='a+')
#     train = train.append(data)
#     del data
elapsed = (time.clock() - start)
print("Time used:",elapsed)

data = pd.read_csv('../data/data_all.csv')

data.head()

data.isnull().sum()

data.info()

bool_col = ['低压开关', '高压开关', '搅拌超压信号', '正泵', '反泵']
data.loc[:, bool_col] = data.loc[:, bool_col].astype('int8')

int_col = ['活塞工作时长', 'sample_file_name']
for i in int_col:
    try:
        data[i] = data[i].astype('int32')
    except:
        data[i] = data[i].astype(str).apply(lambda x:x.split('.')[0]).astype('int32')

data['流量档位'] = data['流量档位'].astype('float32')
label_encoder = data['设备类型'].value_counts().rank(method='dense').astype(int).to_dict()
data['设备类型'] = data['设备类型'].map(label_encoder)
data['设备类型'] = data['设备类型'].astype('int8')

save_variable(data,'../data/data_all.pkl')
save_variable(id_map,'../data/id_map.pkl')
save_variable(label_encoder, '../data/label_encoder.pkl')
