==> code/6_ronghe.ipynb <==
import pandas as pd
from sklearn.metrics import mean_squared_error, f1_score, log_loss,roc_auc_score

"""投票"""
tt = pd.read_csv('../sub/xgb.csv')
dd = pd.read_csv('../sub/cat.csv')
ff = pd.read_csv('../sub/lgb.csv')
tt = tt.merge(dd, on='sample_file_name',how='left')
tt = tt.merge(ff, on='sample_file_name',how='left')
tt['toupiao'] = (tt.label_x+tt.label+tt.label_y)
dic = {0:0,3:1,1:0,2:1}
print(tt['toupiao'].value_counts())
tt['toupiao'] = tt['toupiao'].map(dic)
sub = tt[['sample_file_name', 'toupiao']]
sub.columns = ['sample_file_name', 'label']
print(sub.label.value_counts())
sub.to_csv('../sub/best.csv', index=False)

cc = pd.read_csv('../sub/houchuli.csv')
dd = cc[cc['label'] == 1]['sample_file_name'].values
sub.loc[sub['sample_file_name'].isin(dd), 'label'] = 1
sub.label.value_counts()

sub.to_csv('../sub/sub_3_23.csv', index=False)
